edge-node/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.4.1
eventlet==0.37.0
Pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1
numpy==2.2.1
scipy==1.15.1
